+++ src/Repository/globalQueries.ts
@@ -6,7 +6,7 @@ class GlobalQueries {
         table: string,
         column: string,
         value: string | number | null
-    ): Promise<User[]> {
+    ) {
         const conn = await DB.client.connect();
         const sql = `SELECT * FROM ${table} WHERE ${column}='${value}' ORDER BY id DESC`;
         const res = await conn.query(sql);
@@ -14,7 +14,7 @@ class GlobalQueries {
         return res.rows;
     }
 
-    async updateOne(obj: UpdateOne): Promise<User[]> {
+    async updateOne(obj: UpdateOne) {
         const conn = await DB.client.connect();
         const sql = `UPDATE ${obj.table} SET ${obj.setColumn}='${obj.setValue}' WHERE ${obj.uniqueColumn}='${obj.uniqueValue}' RETURNING *`;
         const res = await conn.query(sql);
