+++ src/models/Mail.ts
@@ -1,0 +1,7 @@
+export interface Mail {
+    email: string;
+    first_name?: string;
+    code?: number;
+    subject: string;
+    message: string;
+}
